+++ lib/td_bg_web/controllers/business_concept_version_controller.ex
@@ -247,7 +247,14 @@ defmodule TdBgWeb.BusinessConceptVersionController do
         |> add_taxonomy()
         |> add_shared_to_parents()
 
-      links = Links.get_links(business_concept_version)
+      links =
+        business_concept_version
+        |> Links.get_links()
+        |> Enum.filter(fn
+          %{domain_id: domain_id} -> can?(claims, view_data_structure(domain_id))
+          _ -> can?(claims, view_data_structure(:no_domain))
+        end)
+
       actions = get_actions(claims, business_concept_version)
 
       shared_to =
@@ -321,7 +328,7 @@ defmodule TdBgWeb.BusinessConceptVersionController do
         }
       }
     else
-        %{}
+      %{}
     end
   end